feat: split programs
