+++ .github/workflows/ci.yml
@@ -73,6 +73,7 @@ jobs:
       - name: Build on-chain programs
         run: |
           cargo build-sbf --manifest-path programs/hydra/Cargo.toml
+          cargo build-sbf --manifest-path programs/hydra-ephemeral/Cargo.toml
           cargo build-sbf --manifest-path tests/programs/noop/Cargo.toml
 
       - name: Test
@@ -85,7 +86,7 @@ jobs:
   # next to this repo.
   # ---------------------------------------------------------------------------
   ephemeral:
-    name: lint + test (ephemeral)
+    name: lint (ephemeral)
     runs-on: ubuntu-latest
     steps:
       - name: Checkout hydra
@@ -115,21 +116,11 @@ jobs:
 
       - name: Clippy — program (ephemeral)
         working-directory: hydra
-        run: cargo clippy -p hydra --features ephemeral --all-targets -- -D warnings
+        run: cargo clippy -p hydra-ephemeral --all-targets -- -D warnings
 
       - name: Clippy — ephemeral test crate
         working-directory: hydra
         run: cargo clippy --manifest-path tests/ephemeral/Cargo.toml --all-targets -- -D warnings
-
-      - name: Build on-chain programs (ephemeral)
-        working-directory: hydra
-        run: |
-          cargo build-sbf --manifest-path programs/hydra/Cargo.toml -- --features ephemeral
-          cargo build-sbf --manifest-path tests/programs/noop/Cargo.toml
-
-      - name: Test — ephemeral lifecycle
-        working-directory: hydra
-        run: cargo nextest run --manifest-path tests/ephemeral/Cargo.toml
 
   # ---------------------------------------------------------------------------
   # Live end-to-end: boots the real three-process stack (mb-test-validator +
@@ -180,7 +171,7 @@ jobs:
 
       - name: Build on-chain programs (ephemeral)
         run: |
-          cargo build-sbf --manifest-path programs/hydra/Cargo.toml -- --features ephemeral
+          cargo build-sbf --manifest-path programs/hydra-ephemeral/Cargo.toml
           cargo build-sbf --manifest-path tests/programs/noop/Cargo.toml
 
       - name: Test — live ephemeral cranks
