+++ programs/hydra/Cargo.toml
@@ -23,20 +23,15 @@ cu-trace = []
 # Uses the SIMD-0312 System Program instruction for prefunded crank PDAs.
 # Keep this disabled on mainnet until `create_account_allow_prefund` is active.
 # NOTE: requires pinocchio-system >= 0.6 (`CreateAccountAllowPrefund`), which is
-# incompatible with the pinocchio 0.10 pin needed for ephemeral support.
+# incompatible with the pinocchio 0.10 pin the ephemeral program needs.
 create-account-allow-prefund = []
-# Ephemeral-rollup crank instructions: schedule/run cranks that live as
-# MagicBlock ephemeral accounts on the ephemeral rollup. Pulls in the
-# `ephemeral-rollups-pinocchio` CPI helpers.
-ephemeral = ["dep:ephemeral-rollups-pinocchio"]
 
 [dependencies]
 pinocchio = { workspace = true }
 pinocchio-system = { workspace = true }
 pinocchio-log = { workspace = true }
-ephemeral-rollups-pinocchio = { workspace = true, optional = true }
 solana-define-syscall = { workspace = true }
-hydra-api = { workspace = true }
+hydra-api = { workspace = true, features = ["program"] }
 
 [package.metadata.solana]
 program-id = "Hydra17i1feui9deaxu6d1TzSQMRNHeBRkDR1Awy7zea"
