--- programs/hydra/src/processor/base/mod.rs
@@ -1,8 +1,0 @@
-//! Base-layer crank lifecycle (feature `base`).
-
-pub mod cancel;
-pub mod close;
-pub mod create;
-pub mod trigger;
-
-mod common;
