+++ programs/hydra/src/processor/cancel.rs
@@ -2,14 +2,14 @@
 
 use pinocchio::{error::ProgramError, AccountView, ProgramResult};
 
-use crate::processor::base::common::drain_lamports;
-use crate::processor::common::require_cancel_authority;
+use crate::processor::common::drain_lamports;
+use hydra_api::program::processor::require_cancel_authority;
 
 pub fn process(accounts: &[AccountView], _data: &[u8]) -> ProgramResult {
     let [authority, crank_ai, recipient] = accounts else {
         return Err(ProgramError::NotEnoughAccountKeys);
     };
 
-    require_cancel_authority(authority, crank_ai, &hydra_api::base::ID)?;
+    require_cancel_authority(authority, crank_ai, &crate::ID)?;
     drain_lamports(crank_ai, recipient)
 }
