--- programs/hydra/src/processor/ephemeral/mod.rs
@@ -1,8 +1,0 @@
-//! Ephemeral-rollup crank lifecycle (feature `ephemeral`).
-
-pub mod cancel;
-pub mod close;
-pub mod create;
-pub mod trigger;
-
-mod common;
